% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualization_functions.R
\name{CreateRPalette}
\alias{CreateRPalette}
\title{Convert Python array of colors from Coolors.co to R-vector of strings}
\usage{
CreateRPalette(array_string, n = 5)
}
\arguments{
\item{array_string}{Python string of color hex codes}

\item{n}{Number of colors in string, default 5}
}
\value{
Vector of hex codes, usable in R
}
\description{
Convert Python array of colors from Coolors.co to R-vector of strings
}
\examples{
coolor_array <- '["845a6d","3e1929","6e75a8","8d91c7","b0daf1"]'
CreateRPalette(coolor_array)
}
