% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{make_latex_table_body}
\alias{make_latex_table_body}
\title{Make the body of a Latex table from INLA object}
\usage{
make_latex_table_body(inla.object)
}
\arguments{
\item{inla.object}{INLA (measurement error) model}
}
\value{
A cleaned INLA parameter summary, with "&" and "\\\\" inserted where necessary to create the body of a LaTeX table, but without the surrounding environment (so that can be specified manually in the manuscript).
}
\description{
Make the body of a Latex table from INLA object
}
