% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_simulation}
\alias{model_simulation}
\title{Example INLA model}
\format{
An INLA model (based on simulated data, with missingness and ME).
}
\usage{
model_simulation
}
\description{
An example INLA model, useful to test functions based on INLA output.
}
\keyword{datasets}
