% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_themes.R
\name{my_basic_theme}
\alias{my_basic_theme}
\title{My basic theme}
\usage{
my_basic_theme(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
a basic ggplot theme with my preferred settings
}
\description{
My basic theme
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + my_basic_theme()
}
