% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{prepare_table}
\alias{prepare_table}
\title{Prepare INLA summary as a clean table}
\usage{
prepare_table(inla.object)
}
\arguments{
\item{inla.object}{INLA (measurement error) model}
}
\value{
A deep cleaned version of the INLA model summary. Rows are given names from the INLA names, and numbers are rounded and padded with zeros when necessary.
}
\description{
Prepare INLA summary as a clean table
}
\examples{
prepare_table(skaRstein::model_simulation)
}
