% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{return_estimates}
\alias{return_estimates}
\title{Return estimates of relevant coefficients from INLA object}
\usage{
return_estimates(inla.object)
}
\arguments{
\item{inla.object}{INLA (measurement error) model}
}
\value{
A table with the posterior means and standard errors (and quantiles) of all estimated parameters in the model whose name contains "beta" or "alpha".
}
\description{
Return estimates of relevant coefficients from INLA object
}
\examples{
return_estimates(skaRstein::model_simulation)
}
