% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualization_functions.R
\name{ShowColors}
\alias{ShowColors}
\title{Show colors}
\usage{
ShowColors(palette)
}
\arguments{
\item{palette}{Vector of hex-codes to be visualized}
}
\value{
ggplot2 object that shows the colors
}
\description{
Show colors
}
\examples{
color_vec <- c("#845a6d", "#3e1929", "#6e75a8", "#8d91c7")
ShowColors(color_vec)
}
